+++ src/main/java/buncheoleasy/user/infrastructure/shipping/ShippingAddressMapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="buncheoleasy.user.infrastructure.ShippingAddressMapper">
+<mapper namespace="buncheoleasy.user.infrastructure.shipping.ShippingAddressMapper">
 
     <!-- ShippingAddress ResultMap -->
     <resultMap id="shippingAddressResultMap" type="ShippingAddress">
@@ -79,13 +79,11 @@
 
     <!-- 중복 배송지 확인 -->
     <select id="existsByUserIdAndShippingMethodAndStoreName" resultType="boolean">
-        SELECT EXISTS(
-                       SELECT 1
-                       FROM shipping_addresses
-                       WHERE user_id = #{userId}
-                         AND shipping_method = #{shippingMethod}
-                         AND store_name = #{storeName}
-                   )
+        SELECT EXISTS(SELECT 1
+                      FROM shipping_addresses
+                      WHERE user_id = #{userId}
+                        AND shipping_method = #{shippingMethod}
+                        AND store_name = #{storeName})
     </select>
 
 </mapper>
